Refactor styled-components to separate files
A significant refactoring has been implemented across multiple components (DefaultPageTemplate, Header, Logo) and pages (RestaurantsPage). Styles initially declared within components have been moved into separate styled-component files for each component. This ensures cleaner code structure with separated concerns between styling and component logic.
