+++ frontend/src/components/Header/Header.styled.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const StyledHeader = styled.header`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 70px;
+    border-bottom: 1px solid var(--default-border-color);
+    border-radius: var(--default-border-radius);
+`;
