+++ frontend/src/components/Logo/Logo.styled.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledLogo = styled.span`
+    font-size: 2rem;
+    font-weight: lighter;
+    font-stretch: expanded;
+`;
