+++ frontend/src/pages/RestaurantsPage.tsx
@@ -1,6 +1,6 @@
 import RestaurantCardList from "../components/RestaurantCardList/RestaurantCardList.tsx";
 import {restaurants} from "../data/restaurantData.ts";
-import DefaultPageTemplate from "./templates/DefaultPageTemplate.tsx";
+import DefaultPageTemplate from "./templates/DefaultPageTemplate/DefaultPageTemplate.tsx";
 
 export default function RestaurantsPage() {
     return (
