+++ frontend/src/pages/templates/DefaultPageTemplate/DefaultPageTemplate.styled.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const StyledMain = styled.main`
+    margin: 20px 20px;
+`;
+
+export const StyledHeading = styled.h1`
+    font-size: 1.7rem;
+    margin-bottom: .7rem;
+`;
